requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
click==6.7
exifread==2.1.2
idna==2.7                 # via requests
imagehash==4.0
numpy==1.15.1             # via imagehash, pywavelets, scipy
pillow==5.2.0
pywavelets==1.0.0         # via imagehash
pyyaml==3.13
requests==2.19.1
scipy==1.1.0              # via imagehash
six==1.11.0               # via imagehash
urllib3==1.23             # via requests
